tidying up after api testing
